requirements.txt
# Core
pytz==2018.3
Django==2.0.2
django-configurations==2.0
gunicorn==19.7.1
newrelic==2.102.0.85

# For the persistence stores
psycopg2==2.7.4
psycopg2-binary==2.7.4
dj-database-url==0.4.2

# Model Tools
django-model-utils==3.1.1
django_unique_upload==0.2.1
django-reversion>=2.0.13,<2.1

# Rest apis
djangorestframework==3.7.7
Markdown==2.6.11
django-filter==1.1.0
django-rest-registration==0.3.7

# Developer Tools
ipdb==0.11
ipython==6.2.1
mkdocs==0.17.2
flake8==3.5.0
django-rest-swagger==2.1.2
django-cors-headers==2.1.0
raven==6.7.0

# Testing
mock==2.0.0
factory-boy==2.10.0
django-nose==1.4.5
nose-progressive==1.5.1
coverage==4.5.1
django-dynamic-fixture==2.0.0

# Static and Media Storage
django-storages==1.6.5
boto3==1.5.31

# External services
django-anymail[mailgun]==1.4
coinmarketcap>=4.2,<4.3
blockchain==1.4
pusher==1.7.4
pywallet>=0.0.7
